Product-Service/requirements.txt
Flask==2.1.2
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.5.1
flask_swagger_ui==3.36.0